Implement CycleMetrics interface for WorkItemDeliveryCycle
[story=PO-92 subject=cycle-metrics-detailed-dashboard story_url=https://urjuna.atlassian.net/browse/PO-92
]

+++ polaris/analytics/service/graphql/interfaces.py
@@ -215,8 +215,8 @@ class DeliveryCycleInfo(graphene.Interface):
 
 
 class CycleMetrics(graphene.Interface):
-    lead_time = graphene.Int(required=False)
-    cycle_time = graphene.Int(required=False)
+    lead_time = graphene.Float(required=False)
+    cycle_time = graphene.Float(required=False)
 
 
 class AggregateCycleMetrics(graphene.Interface):
